requirements.txt
streamlit
easyocr
Pillow
PyMuPDF
openpyxl
pandas
